fix: connect wallet to frieghter issue
